conf: Store cpu count as unsigned int
While we probably won't see machines with more than 65536 cpus for a
while lets store the cpu count as an integer so that we can avoid quite
a lot of overflow checks in our code.

+++ src/conf/domain_conf.h
@@ -2159,8 +2159,8 @@ struct _virDomainDef {
     virDomainBlkiotune blkio;
     virDomainMemtune mem;
 
-    unsigned short vcpus;
-    unsigned short maxvcpus;
+    unsigned int vcpus;
+    unsigned int maxvcpus;
     int placement_mode;
     virBitmapPtr cpumask;
